# server/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "REST API Sandbox Swagger",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://rest-api-sandbox-ds.herokuapp.com/api/v1"
    },
    {
      "url": "http://localhost:8000/api/v1"
    }
  ],
...
  "paths": {
...
    "/posts": {
      "get": {
        "tags": ["Post"],
        "summary": "Get all Posts",
        "operationId": "getAllPosts",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Post"],
        "summary": "Create post",
        "operationId": "createPost",
        "requestBody": {
          "description": "Created post object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {}
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/posts/post/{id}": {
      "get": {
        "tags": ["Post"],
        "summary": "Get single post",
        "operationId": "getSinglePost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of post that need to be returned",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Post"],
        "summary": "Update single post",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateSinglePost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of post that need to be updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated document object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["Post"],
        "summary": "Delete single Post",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteSinglePost",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id of post that need to be deleted",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {}
          },
          "401": {
            "description": "Access denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    }
  },
  "components": {
    "schemas": {
      "Post": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "text": {
            "type": "string"
          }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Basic JWT authorization.",
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}